Fixed suicidal bow. Also renamed some files

+++ source/minepg/rpg_common/rpg/client/gui/GuiKarmaBar.java
@@ -12,6 +12,7 @@ import rpg.lib.Reference;
 import rpg.playerinfo.PlayerInformation;
 
 public class GuiKarmaBar extends Gui {
+
     private Minecraft mc;
 
     public GuiKarmaBar() {
@@ -21,9 +22,7 @@ public class GuiKarmaBar extends Gui {
     @ForgeSubscribe
     public void onRenderHUD(RenderGameOverlayEvent.Pre evt) {
         if (evt.type.equals(RenderGameOverlayEvent.ElementType.HOTBAR)) {
-            ScaledResolution scaledRes = new ScaledResolution(
-                    this.mc.gameSettings, this.mc.displayWidth,
-                    this.mc.displayHeight);
+            ScaledResolution scaledRes = new ScaledResolution(this.mc.gameSettings, this.mc.displayWidth, this.mc.displayHeight);
             int scaledWdt = scaledRes.getScaledWidth() / 2 - (182 / 2);
             int scaledHgt = scaledRes.getScaledHeight() - scaledRes.getScaledHeight() + 6;
 
@@ -34,15 +33,13 @@ public class GuiKarmaBar extends Gui {
             PlayerInformation playerInfo = PlayerInformation.forPlayer(this.mc.thePlayer);
 
             int karmaBarCap = playerInfo.karmaBarCap();
-            
-            if (karmaBarCap > 0)
-            {
+
+            if (karmaBarCap > 0) {
                 short short1 = 182;
-                int currentKarma = (int)(playerInfo.getKarma() * (float)(short1 + 1));
+                int currentKarma = (int) (playerInfo.getKarma() * (short1 + 1));
                 this.drawTexturedModalRect(scaledWdt, scaledHgt, 0, 0, short1, 5);
 
-                if (currentKarma > 0)
-                {
+                if (currentKarma > 0) {
                     this.drawTexturedModalRect(scaledWdt, scaledHgt + 1, 0, 5, currentKarma, 4);
                 }
             }
@@ -50,26 +47,27 @@ public class GuiKarmaBar extends Gui {
 
         this.mc.renderEngine.resetBoundTexture();
     }
-    
+
     public class GuiRenderKarmaLevel extends Gui {
+
         private Minecraft mc;
-        
+
         public GuiRenderKarmaLevel() {
             this.mc = Minecraft.getMinecraft();
         }
-        
+
         @ForgeSubscribe
         public void onRenderKarmaBar(RenderGameOverlayEvent.Text event) {
-            if(event.type.equals(RenderGameOverlayEvent.ElementType.TEXT)) {
+            if (event.type.equals(RenderGameOverlayEvent.ElementType.TEXT)) {
                 ScaledResolution scaledRes = new ScaledResolution(this.mc.gameSettings, this.mc.displayWidth, this.mc.displayHeight);
                 int scaledWdt = scaledRes.getScaledWidth() / 2;
                 int scaledHgt = scaledRes.getScaledHeight() - scaledRes.getScaledHeight() + 1;
-                
+
                 GL11.glColor4f(1.0F, 1.0F, 1.0F, 1.0F);
                 PlayerInformation playerInfo = PlayerInformation.forPlayer(this.mc.thePlayer);
-                
+
                 String karmaLevel = "" + playerInfo.getKarmaLevel();
-                
+
                 this.drawString(this.mc.fontRenderer, karmaLevel, scaledWdt, scaledHgt, 0xFFFFFF);
             }
         }
